feat: added Slack integration + about page
Added a button on the home page to invite yourself to the Mattrax Slack. It uses a Netlify function to do the invite automatically. An about page has been added to give more insight into the project, this also includes a screenshot of the dashboard. NPM deps have been downgraded due to issues with them conflicting with eachother.